+++ include/multi_level_researcher.h
@@ -10,20 +10,20 @@ class MultiLevelSearcher
 private:
     std::vector<HashGrid*> searchers;
     std::vector<std::vector<int>> sortedIndex;
-    std::vector<float> checkedRadiuses;
-    float maxRadius = 0, minRadius = 0, avgRadius = 0;
-    float infFactor;
+    std::vector<double> checkedRadiuses;
+    double maxRadius = 0, minRadius = 0, avgRadius = 0;
+    double infFactor;
 
 public:
-    MultiLevelSearcher(std::vector<Eigen::Vector3f>* particles, float* bounding, std::vector<float>* radiuses, float inf_factor);
+    MultiLevelSearcher(std::vector<Eigen::Vector3d>* particles, double* bounding, std::vector<double>* radiuses, double inf_factor);
     MultiLevelSearcher() {};
     ~MultiLevelSearcher() {};
 
-    inline std::vector<float>* getCheckedRadiuses() {return &checkedRadiuses;}
-    inline float getMaxRadius() {return maxRadius;}
-    inline float getMinRadius() {return minRadius;}
-    inline float getAvgRadius() {return avgRadius;}
+    inline std::vector<double>* getCheckedRadiuses() {return &checkedRadiuses;}
+    inline double getMaxRadius() {return maxRadius;}
+    inline double getMinRadius() {return minRadius;}
+    inline double getAvgRadius() {return avgRadius;}
 
-    void GetNeighbors(const Eigen::Vector3f& pos, std::vector<int>& neighbors);
-    void GetInBoxParticles(const Eigen::Vector3f& box1, const Eigen::Vector3f& box2, std::vector<int>& insides);
+    void GetNeighbors(const Eigen::Vector3d& pos, std::vector<int>& neighbors);
+    void GetInBoxParticles(const Eigen::Vector3d& box1, const Eigen::Vector3d& box2, std::vector<int>& insides);
 };
