+++ src/multi_level_reseacher.cpp
@@ -3,24 +3,24 @@
 
 #include "hash_grid.h"
 
-MultiLevelSearcher::MultiLevelSearcher(std::vector<Eigen::Vector3f>* particles, float* bounding, std::vector<float>* radiuses, float inf_factor)
+MultiLevelSearcher::MultiLevelSearcher(std::vector<Eigen::Vector3d>* particles, double* bounding, std::vector<double>* radiuses, double inf_factor)
 {
-    std::vector<std::vector<Eigen::Vector3f>> sortedParticles;
+    std::vector<std::vector<Eigen::Vector3d>> sortedParticles;
     maxRadius = *std::max_element(radiuses->begin(), radiuses->end());
     minRadius = *std::min_element(radiuses->begin(), radiuses->end());
     infFactor = inf_factor;
-    std::vector<std::pair<float, float>> bin_bounds;
-    float bin_extent = minRadius * 0.5;
+    std::vector<std::pair<double, double>> bin_bounds;
+    double bin_extent = minRadius * 0.5;
     int bins = std::max(int(ceil((maxRadius - minRadius) / bin_extent)), 1);
     bin_extent = (maxRadius - minRadius) / bins;
     for (size_t i = 0; i < bins; i++)
     {
         bin_bounds.push_back(
-            std::pair<float, float>(
+            std::pair<double, double>(
                 minRadius+(i*bin_extent), 
                 std::min(minRadius+((i+1)*bin_extent), maxRadius)));
     }
-    auto whichBin = [&](const float r)
+    auto whichBin = [&](const double r)
     {
         for (auto tit = bin_bounds.begin(); tit < bin_bounds.end(); tit++)
         {
@@ -50,7 +50,7 @@ MultiLevelSearcher::MultiLevelSearcher(std::vector<Eigen::Vector3f>* particles, 
     if (avgRadius > maxRadius) avgRadius = maxRadius;
     for (int i = 0; i < sortedParticles.size(); i++)
     {
-        float temp_bounding [6] = {0.0f};
+        double temp_bounding [6] = {0.0f};
         temp_bounding[0] = bounding[0];
         temp_bounding[1] = bounding[1];
         temp_bounding[2] = bounding[2];
@@ -59,12 +59,12 @@ MultiLevelSearcher::MultiLevelSearcher(std::vector<Eigen::Vector3f>* particles, 
         temp_bounding[5] = bounding[5];
         searchers.push_back(new HashGrid(sortedParticles[i], temp_bounding, bin_bounds[i].second, inf_factor));
     }
-    std::set<float> st(radiuses->begin(), radiuses->end());
+    std::set<double> st(radiuses->begin(), radiuses->end());
     checkedRadiuses.assign(st.begin(), st.end());
     printf("   Seachers level: %d.\n", searchers.size());
 }
 
-void MultiLevelSearcher::GetNeighbors(const Eigen::Vector3f& pos, std::vector<int>& neighbors)
+void MultiLevelSearcher::GetNeighbors(const Eigen::Vector3d& pos, std::vector<int>& neighbors)
 {
     std::vector<int> subNeighbors;
     for (size_t searcherId = 0; searcherId < searchers.size(); searcherId++)
@@ -78,7 +78,7 @@ void MultiLevelSearcher::GetNeighbors(const Eigen::Vector3f& pos, std::vector<in
     }
 }
 
-void MultiLevelSearcher::GetInBoxParticles(const Eigen::Vector3f& box1, const Eigen::Vector3f& box2, std::vector<int>& insides)
+void MultiLevelSearcher::GetInBoxParticles(const Eigen::Vector3d& box1, const Eigen::Vector3d& box2, std::vector<int>& insides)
 {
     std::vector<int> subInsides;
     for (size_t searcherId = 0; searcherId < searchers.size(); searcherId++)
